fork update @ 20.03.2020

+++ Data/MySQL/src/StatementExecutor.cpp
@@ -1,8 +1,6 @@
 //
 // StatementExecutor.cpp
 //
-// $Id: //poco/1.3/Data/MySQL/src/StatementExecutor.cpp#1 $
-//
 // Library: Data/MySQL
 // Package: MySQL
 // Module:  StatementExecutor
@@ -141,10 +139,11 @@ bool StatementExecutor::fetchColumn(std::size_t n, MYSQL_BIND *bind)
 	return (res == 0);
 }
 
+
 int StatementExecutor::getAffectedRowCount() const
 {
-	return _affectedRowCount;
+	return static_cast<int>(_affectedRowCount);
 }
 
 
-}}}
+} } } // namespace Poco::Data::MySQL
